+++ docs/whats_new.rst
@@ -24,6 +24,7 @@ New Features
 - Add :class:`flogin.jsonrpc.errors.FlowError`
 - Add the ability to use :meth:`flogin.plugin.Plugin.search` as a classmethod
 - Allow :meth:`flogin.jsonrpc.results.Result.callback` to return :class:`bool` and ``None`` ontop of :class:`flogin.jsonrpc.responses.ExecuteResponse`
+- Add the ``name`` argument in the :func:`flogin.utils.print` function.
 
 Bug Fixes
 ~~~~~~~~~